增加使用RecyclerView实现Banner的演示
