+++ app/src/main/java/com/xuexiang/xuidemo/adapter/RecyclerViewBannerAdapter.java
@@ -1,0 +1,157 @@
+/*
+ * Copyright (C) 2019 xuexiangjys(xuexiangjys@163.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.xuexiang.xuidemo.adapter;
+
+import android.graphics.Color;
+import android.graphics.drawable.ColorDrawable;
+import android.text.TextUtils;
+import android.view.View;
+import android.widget.ImageView;
+
+import com.bumptech.glide.load.engine.DiskCacheStrategy;
+import com.xuexiang.xui.widget.banner.recycler.BannerLayout;
+import com.xuexiang.xui.widget.banner.widget.banner.base.GlideImageLoader;
+import com.xuexiang.xui.widget.banner.widget.banner.base.ImageLoader;
+import com.xuexiang.xuidemo.R;
+import com.xuexiang.xuidemo.adapter.base.BaseRecyclerAdapter;
+import com.xuexiang.xuidemo.adapter.base.RecyclerViewHolder;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * @author xuexiang
+ * @since 2019-05-30 00:50
+ */
+public class RecyclerViewBannerAdapter extends BaseRecyclerAdapter<String> {
+
+    /**
+     * 默认加载图片
+     */
+    private ColorDrawable mColorDrawable;
+
+    /**
+     * 是否允许进行缓存
+     */
+    private boolean mEnableCache = true;
+
+    private ImageLoader mImageLoader;
+
+    private BannerLayout.OnBannerItemClickListener mOnBannerItemClickListener;
+
+
+    public RecyclerViewBannerAdapter() {
+        super();
+        mColorDrawable = new ColorDrawable(Color.parseColor("#555555"));
+    }
+
+    public RecyclerViewBannerAdapter(List<String> list) {
+        super(list);
+        mColorDrawable = new ColorDrawable(Color.parseColor("#555555"));
+    }
+
+    public RecyclerViewBannerAdapter(String[] list) {
+        super(Arrays.asList(list));
+        mColorDrawable = new ColorDrawable(Color.parseColor("#555555"));
+    }
+    /**
+     * 适配的布局
+     *
+     * @param viewType
+     * @return
+     */
+    @Override
+    public int getItemLayoutId(int viewType) {
+        return R.layout.adapter_recycler_view_banner_image_item;
+    }
+
+    /**
+     * 绑定数据
+     *
+     * @param holder
+     * @param position
+     * @param imgUrl
+     */
+    @Override
+    public void bindData(RecyclerViewHolder holder, final int position, String imgUrl) {
+        ImageView imageView = holder.findViewById(R.id.iv_item);
+
+        if (!TextUtils.isEmpty(imgUrl)) {
+            getImageLoader().displayImage(imageView.getContext(), imgUrl, imageView, mColorDrawable,
+                    mEnableCache ? DiskCacheStrategy.RESOURCE : DiskCacheStrategy.NONE);
+        } else {
+            imageView.setImageDrawable(mColorDrawable);
+        }
+
+        imageView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (mOnBannerItemClickListener != null) {
+                    mOnBannerItemClickListener.onItemClick(position);
+                }
+            }
+        });
+    }
+
+    private ImageLoader getImageLoader() {
+        if (mImageLoader == null) {
+            mImageLoader = new GlideImageLoader();
+        }
+        return  mImageLoader;
+    }
+
+
+    /**
+     * 设置是否允许缓存
+     *
+     * @param enableCache
+     * @return
+     */
+    public RecyclerViewBannerAdapter enableCache(boolean enableCache) {
+        mEnableCache = enableCache;
+        return this;
+    }
+
+    /**
+     * 获取是否允许缓存
+     *
+     * @return
+     */
+    public boolean getEnableCache() {
+        return mEnableCache;
+    }
+
+    public ColorDrawable getColorDrawable() {
+        return mColorDrawable;
+    }
+
+    public RecyclerViewBannerAdapter setColorDrawable(ColorDrawable colorDrawable) {
+        mColorDrawable = colorDrawable;
+        return this;
+    }
+
+    public RecyclerViewBannerAdapter setImageLoader(ImageLoader imageLoader) {
+        mImageLoader = imageLoader;
+        return this;
+    }
+
+    public RecyclerViewBannerAdapter setOnBannerItemClickListener(BannerLayout.OnBannerItemClickListener onBannerItemClickListener) {
+        mOnBannerItemClickListener = onBannerItemClickListener;
+        return this;
+    }
+}
