+++ app/src/main/res/layout/fragment_recyclerview_banner.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (C) 2019 xuexiangjys(xuexiangjys@163.com)
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~       http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+
+    <com.xuexiang.xui.widget.banner.recycler.BannerLayout
+        android:id="@+id/bl_horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        app:bl_autoPlaying="true"
+        app:bl_centerScale="1.3"
+        app:bl_itemSpace="20"
+        app:bl_moveSpeed="1.8" />
+
+    <com.xuexiang.xui.widget.banner.recycler.BannerLayout
+        android:id="@+id/bl_vertical"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:layout_marginTop="10dp"
+        app:bl_orientation="vertical"
+        app:bl_showIndicator="false" />
+
+</LinearLayout>
