+++ app/src/main/res/layout/fragment_view_pager_banner.xml
