+++ xui_lib/src/main/java/com/xuexiang/xui/widget/banner/widget/loopviewpager/LoopViewPager.java
@@ -75,6 +75,7 @@ public class LoopViewPager extends ViewPager {
         return mAdapter != null ? mAdapter.toRealPosition(super.getCurrentItem()) : 0;
     }
 
+    @Override
     public void setCurrentItem(int item, boolean smoothScroll) {
         int realItem = mAdapter.toInnerPosition(item);
         super.setCurrentItem(realItem, smoothScroll);
